Project hierarchy refactoring (#14)
