+++ src/EasyDialog/EasyDialog.csproj
@@ -63,32 +63,31 @@
     <Compile Include="Enums\MaterialThemeAccent.cs" />
     <Compile Include="Enums\MaterialThemePrimaryColor.cs" />
     <Compile Include="Enums\MetroTheme.cs" />
-    <Compile Include="Internal\Forms\EasyDialogForm.cs" />
-    <Compile Include="Internal\Forms\Interfaces\IFormProvider.cs" />
-    <Compile Include="Internal\Forms\Templates\ButtonTemplate.cs" />
+    <Compile Include="Internal\EasyDialogForm.cs" />
+    <Compile Include="Internal\Providers\IFormProvider.cs" />
+    <Compile Include="Internal\Templates.cs" />
     <Compile Include="Models\MaterialColorScheme.cs" />
-    <Compile Include="Items\ComboBoxItemOptionsBuilder.cs" />
-    <Compile Include="Items\DateTimePickerItemOptionsBuilder.cs" />
-    <Compile Include="Items\Implementations\ComboBoxItem.cs" />
-    <Compile Include="Items\Implementations\DateTimePickerItem.cs" />
-    <Compile Include="Items\Implementations\ListBoxItem.cs" />
-    <Compile Include="Items\Implementations\NumericUpDownItem.cs" />
-    <Compile Include="Items\Implementations\CheckBoxItem.cs" />
-    <Compile Include="Items\CheckBoxItemOptionsBuilder.cs" />
-    <Compile Include="Items\ListBoxItemOptionsBuilder.cs" />
-    <Compile Include="Items\NumericUpDownItemOptionsBuilder.cs" />
-    <Compile Include="Items\TextBoxItemOptionsBuilder.cs" />
-    <Compile Include="Items\DialogItemOptionsBuilder.cs" />
+    <Compile Include="Items\OptionsBuilders\ComboBoxItemOptionsBuilder.cs" />
+    <Compile Include="Items\OptionsBuilders\DateTimePickerItemOptionsBuilder.cs" />
+    <Compile Include="Items\ComboBoxItem.cs" />
+    <Compile Include="Items\DateTimePickerItem.cs" />
+    <Compile Include="Items\ListBoxItem.cs" />
+    <Compile Include="Items\NumericUpDownItem.cs" />
+    <Compile Include="Items\CheckBoxItem.cs" />
+    <Compile Include="Items\OptionsBuilders\CheckBoxItemOptionsBuilder.cs" />
+    <Compile Include="Items\OptionsBuilders\ListBoxItemOptionsBuilder.cs" />
+    <Compile Include="Items\OptionsBuilders\NumericUpDownItemOptionsBuilder.cs" />
+    <Compile Include="Items\OptionsBuilders\TextBoxItemOptionsBuilder.cs" />
+    <Compile Include="Items\OptionsBuilders\DialogItemOptionsBuilder.cs" />
     <Compile Include="Enums\DialogStyle.cs" />
-    <Compile Include="Internal\Forms\Implementations\DefaultFormProvider.cs" />
-    <Compile Include="Internal\Forms\DialogContextFormLoader.cs" />
-    <Compile Include="Internal\Forms\Implementations\MaterialFormProvider.cs" />
-    <Compile Include="Internal\Forms\Implementations\MetroFormProvider.cs" />
-    <Compile Include="Internal\Forms\Interfaces\IEasyDialogForm.cs" />
-    <Compile Include="Internal\Forms\Templates\LabelTemplate.cs" />
-    <Compile Include="Items\Interfaces\BaseDialogItem.cs" />
-    <Compile Include="Internal\Items\DialogContextItemsLoader.cs" />
-    <Compile Include="Items\Implementations\TextBoxItem.cs" />
+    <Compile Include="Internal\Providers\DefaultFormProvider.cs" />
+    <Compile Include="Internal\DialogContextFormLoader.cs" />
+    <Compile Include="Internal\Providers\MaterialFormProvider.cs" />
+    <Compile Include="Internal\Providers\MetroFormProvider.cs" />
+    <Compile Include="Internal\IEasyDialogForm.cs" />
+    <Compile Include="Items\BaseDialogItem.cs" />
+    <Compile Include="Internal\DialogContextItemsLoader.cs" />
+    <Compile Include="Items\TextBoxItem.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup />
