--- src/EasyDialog/Internal/Forms/Templates/ButtonTemplate.cs
@@ -1,9 +1,0 @@
-﻿using System.Windows.Forms;
-
-namespace bubuntoid.EasyDialog.Internal.Forms.Templates
-{
-    internal static class ButtonTemplate
-    {
-        public static Button DefaultButton => new Button();
-    }
-}
