+++ src/EasyDialog/Items/BaseDialogItem.cs
