+++ src/EasyDialog/Items/CheckBoxItem.cs
