+++ src/EasyDialog/Items/ListBoxItem.cs
