+++ src/EasyDialog/Items/NumericUpDownItem.cs
