+++ src/EasyDialog/Items/OptionsBuilders/CheckBoxItemOptionsBuilder.cs
