+++ src/EasyDialog/Items/OptionsBuilders/NumericUpDownItemOptionsBuilder.cs
