+++ src/EasyDialog/Items/TextBoxItem.cs
